refactored fileUtils
