+++ app/src/main/res/values-pl/strings.xml
@@ -111,6 +111,7 @@
     <string name="label_show_more_details">Pokaż więcej szczegółów</string>
     <string name="label_menu_item_management">Zarzadzanie pozycją menu</string>
     <string name="label_menu_item_photo">Zdjęcie pozycji w menu</string>
+    <string name="label_menu_photo">Zdjęcie menu</string>
     <string name="label_edit_menu_item">Edytuj pozycję w menu</string>
     <string name="label_price">Cena</string>
     <string name="label_alcohol">Zawartość alkoholu</string>
@@ -124,4 +125,6 @@
     <string name="label_stats">Statystyki</string>
     <string name="label_employees">Pracownicy</string>
     <string name="label_delete">Usuń</string>
+    <string name="error_file_not_given">Plik jest wymagany</string>
+    <string name="error_wrong_file_format">Zły format pliku</string>
 </resources>
