+++ app/src/main/res/values/strings.xml
@@ -116,6 +116,7 @@
     <string name="label_show_more_details">Show more details</string>
     <string name="label_menu_item_management">Menu item management</string>
     <string name="label_menu_item_photo">Menu Item Photo</string>
+    <string name="label_menu_photo">Zdjęcie pozycji z menu</string>
     <string name="label_edit_menu_item">Edit menu item</string>
     <string name="label_price">Price</string>
     <string name="label_alcohol">Alcohol Percentage</string>
@@ -129,4 +130,6 @@
     <string name="label_stats">Stats</string>
     <string name="label_employees">Employees</string>
     <string name="label_delete">Delete</string>
+    <string name="error_file_not_given">File is required</string>
+    <string name="error_wrong_file_format">Wrong file format</string>
 </resources>
